Add support for OpenID extension.
This work originates from the work of @jeffmiller at https://github.com/mitsuhiko/flask-openid/pull/10
I added on the top of that the possibility to parse the OpenID response with the OpenID
extension of interest.

With this patch one can use flask-openid with an OpenID supporting the team or the CLA
extensions.